chat-users/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-crispy-forms==1.7.2
django-filebrowser==3.11.1
django-filter==2.2.0
django-filters==0.2.1
django-grappelli==2.12.3
django-summernote==0.8.11.4
django-templated-mail==1.1.1
django-thumbs-v2==0.4.1
djangorestframework==3.10.2
djoser==2.0.1
idna==2.8
Markdown==3.1.1
nodeenv==1.3.3
npm==0.1.1
optional-django==0.1.0
Pillow==6.1.0
psycopg2==2.8.3
python-resize-image==1.1.19
pytz==2019.2
requests==2.22.0
selenium==3.141.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
urllib3==1.25.3
